# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/commissions/runs": {
      "get": {
        "operationId": "CommissionsAdminController_listRuns",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "IN_PROGRESS",
                "SUCCESS",
                "ERROR"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Runs exécutés à partir de cette date (incluse), au format ISO.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Runs exécutés jusqu’à cette date (incluse), au format ISO.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPageDto"
                }
              }
            }
          }
        },
        "summary": "Historique des runs hebdomadaires (période, membres réglés, total distribué, statut).",
        "tags": [
          "commissions-admin"
        ]
      },
      "post": {
        "operationId": "CommissionsAdminController_relaunch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelaunchRunDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "memberCount": {
                      "type": "number"
                    },
                    "distributedDt": {
                      "type": "string",
                      "example": "18750.000"
                    },
                    "rewardPointsGranted": {
                      "type": "number"
                    },
                    "eventCount": {
                      "type": "number"
                    },
                    "alreadyExecuted": {
                      "type": "boolean"
                    },
                    "periodStart": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "periodEnd": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Relance de secours d’un run (SUPER_ADMIN). Idempotente : ne recrédite jamais une période déjà réglée.",
        "tags": [
          "commissions-admin"
        ]
      }
    },
...
    "/admin/commissions/runs/{runId}": {
      "get": {
        "operationId": "CommissionsAdminController_runDetail",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailDto"
                }
              }
            }
          }
        },
        "summary": "Détail d’un run : totaux, argent PERDU au plafond, événements inéligibles, journal d’exécution.",
        "tags": [
          "commissions-admin"
        ]
      }
    },
    "/admin/commissions/runs/{runId}/members": {
      "get": {
        "operationId": "CommissionsAdminController_runMembers",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunMemberPageDto"
                }
              }
            }
          }
        },
        "summary": "Décomposition par membre d’un run : brut, versé, perdu au plafond, plafond appliqué, Points Fidélité.",
        "tags": [
          "commissions-admin"
        ]
      }
    },
    "/admin/commissions/runs/{runId}/members/{memberId}/events": {
      "get": {
        "description": "La part payée et la part perdue de CHAQUE événement ne sont pas stockées : elles sont rejouées par la fonction de règlement du moteur lui-même, sur les mêmes entrées. C’est l’explication exacte du versement, pas une reconstitution.",
        "operationId": "CommissionsAdminController_memberEvents",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunMemberEventsDto"
                }
              }
            }
          }
        },
        "summary": "Chronologie des événements d’un membre sur un run, dans l’ordre d’application du plafond (D-033).",
        "tags": [
          "commissions-admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Najah Partners API",
    "description": "API backend — plateforme MLM (BV, e-cards, commissions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RunSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "executedAt": {
            "format": "date-time",
            "type": "string"
          },
          "periodStart": {
            "format": "date-time",
            "type": "string",
            "description": "Début INCLUS de la semaine réglée."
          },
          "periodEnd": {
            "format": "date-time",
            "type": "string",
            "description": "Fin EXCLUE (clôture vendredi 23:59 Tunis)."
          },
          "memberCount": {
            "type": "number",
            "description": "Membres ayant reçu un versement."
          },
          "distributedDt": {
            "type": "string",
            "example": "18750.000",
            "description": "DINARS versés par ce run."
          },
          "rewardPointsGranted": {
            "type": "number",
            "description": "Points Fidélité accordés (D-032)."
          },
          "status": {
            "enum": [
              "IN_PROGRESS",
              "SUCCESS",
              "ERROR"
            ],
            "type": "string"
          },
          "hasLog": {
            "type": "boolean",
            "description": "Le run porte-t-il un journal d’exécution ?"
          }
        },
        "required": [
          "id",
          "executedAt",
          "periodStart",
          "periodEnd",
          "memberCount",
          "distributedDt",
          "rewardPointsGranted",
          "status",
          "hasLog"
        ]
      },
      "RunPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunSummaryDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
...
      "RunDetailDto": {
        "type": "object",
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunSummaryDto"
          },
          "log": {
            "type": "string",
            "nullable": true,
            "description": "Journal d’exécution (texte libre écrit par le run, `null` s’il n’y en a pas)."
          },
          "grossTotalDt": {
            "type": "string",
            "example": "19750.000",
            "description": "DINARS — somme des bruts de tous les membres."
          },
          "lostTotalDt": {
            "type": "string",
            "example": "1000.000",
            "description": "DINARS perdus au plafond sur l’ensemble du run."
          },
          "rewardPointsLost": {
            "type": "number",
            "description": "Points Fidélité perdus au plafond sur l’ensemble du run."
          },
          "eventCount": {
            "type": "number",
            "description": "Événements réclamés par ce run (éligibles et inéligibles)."
          },
          "ineligibleEventCount": {
            "type": "number",
            "description": "Événements réclamés mais INÉLIGIBLES (bénéficiaire gelé ou INSCRIT — D-034) : soldés, jamais payés."
          },
          "ineligibleGrossDt": {
            "type": "string",
            "example": "500.000",
            "description": "DINARS que ces événements inéligibles auraient valus."
          },
          "unsettledMemberCount": {
            "type": "number",
            "description": "Bénéficiaires n’ayant reçu AUCUN règlement (tous leurs événements étaient inéligibles) — ils n’apparaissent pas dans la décomposition par membre."
          }
        },
        "required": [
          "run",
          "grossTotalDt",
          "lostTotalDt",
          "rewardPointsLost",
          "eventCount",
          "ineligibleEventCount",
          "ineligibleGrossDt",
          "unsettledMemberCount"
        ]
      },
      "RunMemberRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "memberCode": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "enum": [
              "REGISTERED",
              "ACTIVE",
              "INACTIVE"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "memberCode",
          "firstName",
          "lastName",
          "status"
        ]
      },
      "RunMemberRowDto": {
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/RunMemberRefDto"
          },
          "grossDt": {
            "type": "string",
            "example": "11000.000",
            "description": "DINARS — total éligible AVANT plafond."
          },
          "paidDt": {
            "type": "string",
            "example": "10000.000",
            "description": "DINARS réellement versés."
          },
          "lostDt": {
            "type": "string",
            "example": "1000.000",
            "description": "DINARS PERDUS au plafond — jamais reportés (D-033). = brut − versé."
          },
          "appliedCapDt": {
            "type": "string",
            "example": "10000.000",
            "description": "Plafond appliqué, issu du SNAPSHOT d’activation du membre — jamais du pack vivant."
          },
          "eventCount": {
            "type": "number"
          },
          "rewardPointsGranted": {
            "type": "number",
            "description": "Points Fidélité accordés sous le plafond."
          },
          "rewardPointsLost": {
            "type": "number",
            "description": "Points Fidélité perdus (survenus après le plafond)."
          }
        },
        "required": [
          "member",
          "grossDt",
          "paidDt",
          "lostDt",
          "appliedCapDt",
          "eventCount",
          "rewardPointsGranted",
          "rewardPointsLost"
        ]
      },
      "RunMemberPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunMemberRowDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "RunEventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "enum": [
              "DIRECT",
              "BALANCE",
              "STARTUP_BONUS",
              "REWARD_POINT"
            ],
            "type": "string"
          },
          "amountDt": {
            "type": "string",
            "example": "250.000",
            "description": "DINARS dus par l’événement — toujours 0 pour un REWARD_POINT (D-032)."
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string",
            "description": "Horodatage RÉEL de la transaction d’activation (porte la chronologie)."
          },
          "sourceMember": {
            "description": "Le filleul dont l’activation a produit l’événement.",
            "allOf": [
              {
                "$ref": "#/components/schemas/RunMemberRefDto"
              }
            ]
          },
          "eligible": {
            "type": "boolean",
            "description": "Éligibilité évaluée AU MOMENT de l’événement (D-034) : un bénéficiaire gelé ou INSCRIT donne `false` — tracé, jamais payé."
          },
          "balanceIndex": {
            "type": "number",
            "nullable": true,
            "description": "N° d’équilibre À VIE (D-032) — `null` pour une commission directe."
          },
          "cumulativeBeforeDt": {
            "type": "string",
            "example": "9800.000",
            "description": "DINARS — cumul de la semaine AVANT cet événement."
          },
          "paidDt": {
            "type": "string",
            "example": "200.000",
            "description": "DINARS effectivement versés par cet événement."
          },
          "lostDt": {
            "type": "string",
            "example": "50.000",
            "description": "DINARS perdus AU PLAFOND sur cet événement (D-033). Vaut 0 pour un événement INÉLIGIBLE : cette somme n’a jamais été due — c’est `eligible` qui le dit, pas ce montant."
          },
          "crossesCap": {
            "type": "boolean",
            "description": "Cet événement FRANCHIT le plafond : il est payé partiellement."
          },
          "rewardPointGranted": {
            "type": "boolean",
            "description": "Point Fidélité accordé par cet événement."
          },
          "rewardPointLost": {
            "type": "boolean",
            "description": "Point Fidélité perdu (événement survenu après le plafond)."
          }
        },
        "required": [
          "id",
          "type",
          "amountDt",
          "occurredAt",
          "sourceMember",
          "eligible",
          "cumulativeBeforeDt",
          "paidDt",
          "lostDt",
          "crossesCap",
          "rewardPointGranted",
          "rewardPointLost"
        ]
      },
      "RunMemberEventsDto": {
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/RunMemberRefDto"
          },
          "appliedCapDt": {
            "type": "string",
            "nullable": true,
            "example": "10000.000",
            "description": "Plafond appliqué à cette semaine (snapshot d’activation). `null` si aucun règlement n’a eu lieu — tous les événements étaient inéligibles, le plafond n’a jamais eu à s’appliquer."
          },
          "grossDt": {
            "type": "string",
            "example": "11000.000"
          },
          "paidDt": {
            "type": "string",
            "example": "10000.000"
          },
          "lostDt": {
            "type": "string",
            "example": "1000.000"
          },
          "events": {
            "description": "Ordre chronologique STRICT `(occurredAt, id)` — l’ordre même d’application du plafond.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunEventDto"
            }
          }
        },
        "required": [
          "member",
          "grossDt",
          "paidDt",
          "lostDt",
          "events"
        ]
      },
      "RelaunchRunDto": {
        "type": "object",
        "properties": {
          "periodEnd": {
            "type": "string",
            "description": "Clôture (fin EXCLUE) de la semaine à régler, au format ISO. Omise : la dernière semaine close."
          }
        }
      },
...
    }
  }
}